Implemented KodeKsData.laserFields
This is a LaserField ArrayList that holds all lasers to be drawn.

laserFields works similarly to the list of threatened stones (that badly
needs a name change) but in addition includes an orientation, since
lasers can be oriented horizontally, vertically, diagonally from NW to
SE (diagonalNW) or diagonally from NE to SW (diagonalNE).

Next Task:
Make use of this during drawing.